test: Silence warnings on Win64 builds

+++ test/bio_base64_test.c
@@ -100,7 +100,7 @@ static int encode(unsigned const char *buf, unsigned buflen, char *encoded,
 
     /* Use a verbatim encoding when provided */
     if (encoded != NULL) {
-        int elen = strlen(encoded);
+        int elen = (int)strlen(encoded);
 
         return BIO_write(mem, encoded, elen) == elen;
     }
@@ -152,8 +152,8 @@ static int genb64(char *prefix, char *suffix, unsigned const char *buf,
                   unsigned buflen, int trunc, char *encoded, unsigned llen,
                   unsigned wscnt, char **out)
 {
-    int preflen = strlen(prefix);
-    int sufflen = strlen(suffix);
+    int preflen = (int)strlen(prefix);
+    int sufflen = (int)strlen(suffix);
     int outlen;
     char newline = '\n';
     BUF_MEM *bptr;
@@ -174,7 +174,7 @@ static int genb64(char *prefix, char *suffix, unsigned const char *buf,
     /* Orphan the memory BIO's data buffer */
     BIO_get_mem_ptr(mem, &bptr);
     *out = bptr->data;
-    outlen = bptr->length;
+    outlen = (int)bptr->length;
     bptr->data = NULL;
     (void) BIO_set_close(mem, BIO_NOCLOSE);
     BIO_free(mem);
